refactor: replace 558KB png placeholder with inline svg fallback asset ss-111
